requirements.txt
kafka-python==2.0.2
redis==4.5.4
pymilvus==2.3.3
numpy>=1.20.0
pandas>=1.4.0
requests==2.31.0
torch>=2.0.0
torch-geometric
scikit-learn